added some changes to the goals chart
